social_media_api/requirements.txt
asgiref==3.8.1
autopep8==2.3.2
certifi==2025.1.31
distlib==0.3.9
Django==5.1.6
django-filter==25.1
django-taggit==6.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.9
filelock==3.17.0
inflection==0.5.1
Markdown==3.7
mysqlclient==2.2.7
packaging==24.2
pillow==11.1.0
pipenv==2024.4.1
platformdirs==4.3.6
pycodestyle==2.12.1
PyJWT==2.9.0
pytz==2025.1
PyYAML==6.0.2
setuptools==75.8.2
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
virtualenv==20.29.2
